--- a/src/Hl7.FhirPath.Tests/Tests/FhirPathTest.cs
+++ b/src/Hl7.FhirPath.Tests/Tests/FhirPathTest.cs
@@ -17,12 +17,20 @@ using Hl7.Fhir.Model.Primitives;
 using Hl7.Fhir.Model;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Hl7.Fhir.FhirPath;
+using System.Diagnostics;
+using System.Threading.Tasks.Dataflow;
 
 namespace Hl7.FhirPath.Tests
 {
     [TestClass]
     public class FhirPathTest
     {
+        /// <summary>
+        ///  Gets or sets the test context which provides
+        ///  information about and functionality for the current test run.
+        ///</summary>
+        public TestContext TestContext { get; set; }
+
         [TestMethod]
         public void ConvertToInteger()
         {
@@ -171,6 +179,42 @@ namespace Hl7.FhirPath.Tests
 
             result = patient.Predicate("identifier[0].value.combine($this.identifier[0].value).isDistinct()");
             Assert.IsFalse(result);
+        }
+
+        [TestMethod]
+        [TestCategory("LongRunner")]
+        public void TestFhirPathScalarInParallel()
+        {
+            var patient = new Patient();
+
+            patient.Identifier.Add(new Identifier("http://nu.nl", "id1"));
+
+            var element = patient.ToTypedElement();
+
+            Stopwatch stopwatch = new Stopwatch();
+            stopwatch.Start();
+
+            var actionBlock = new ActionBlock<string>(
+                 (input =>
+                 {
+                     element.Select(input);
+                 }),
+                 new ExecutionDataflowBlockOptions()
+                 {
+                     MaxDegreeOfParallelism = 20
+                 });
+
+
+            for (int i = 0; i < 20_000; i++)
+            {
+                actionBlock.Post($"identifier[{i}]");
+            }
+            actionBlock.Complete();
+            actionBlock.Completion.Wait();
+
+            stopwatch.Stop();
+
+            TestContext.WriteLine($"Select in 20.000 : {stopwatch.ElapsedMilliseconds}ms");
         }
 
         //[TestMethod]