Solves issue #561.

Time measurements:

_Before this improvement:_
- hitting the cache 20,000 times in 20 parallel tasks took an average time of 7,850ms (almost 8 sec).

This improvement:
- hitting the cache 20,000 times in 20 parallel tasks took an average time of 5,691ms.
- when we empty the cache by 10% of its size we get even a better performance. Then we have 4,725ms of time. This method is now implemented.