@ewoutkramer 2018-12-13
Fixes #561